requirements.txt
openai
transformers
optimum
pinecone-client
accelerate
fastapi
uvicorn[standard]